pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "python-lamp-web-app"
version = "0.1.0"
description = "An enhanced interactive lamp web application with 3D animations and effects"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Jared T"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.116.1",
    "uvicorn[standard]==0.35.0",
    "jinja2==3.1.6",
    "python-multipart==0.0.20",
    "pydantic==2.11.7",
    # Database dependencies
    "sqlalchemy==2.0.36",
    "pyodbc==5.2.0",
    "alembic==1.14.0",
    # Azure SDK for Key Vault
    "azure-keyvault-secrets==4.8.0",
    "azure-identity==1.19.0",
    # Additional utilities
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/jaredthivener/python-lamp-web-app"
Repository = "https://github.com/jaredthivener/python-lamp-web-app"
Issues = "https://github.com/jaredthivener/python-lamp-web-app/issues"

[project.scripts]
lamp-app = "src.main:main"

# Hatchling build configuration
[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
]

[tool.black]
target-version = ['py312']
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
rules = [
    "auto",  # Auto-detect language and apply relevant rules
    "security",
    "python",
]
